cap2-django/section01-first_course/django_project/src/app/router.py
from rest_framework import routers
from .viewsets import UserViewSet, TagViewSet, PostViewSet, AccountViewSet

router = routers.DefaultRouter()

router.register('users', UserViewSet)
router.register('tags', TagViewSet)
router.register('posts', PostViewSet)
router.register('accounts', AccountViewSet)
